Move matcher services to an extension
