+++ src/PhpSpec/Console/ContainerAssembler.php
@@ -13,7 +13,8 @@
 
 namespace PhpSpec\Console;
 
-use PhpSpec\IO\IOExtension;
+use PhpSpec\IO\DependencyInjection\IOExtension;
+use PhpSpec\Matcher\DependencyInjection\MatcherExtension;
 use SebastianBergmann\Exporter\Exporter;
 use PhpSpec\Process\ReRunner;
 use PhpSpec\Util\MethodAnalyser;
@@ -24,7 +25,6 @@ use PhpSpec\Formatter as SpecFormatter;
 use PhpSpec\Listener;
 use PhpSpec\Loader;
 use PhpSpec\Locator;
-use PhpSpec\Matcher;
 use PhpSpec\Runner;
 use PhpSpec\Wrapper;
 use RuntimeException;
@@ -432,42 +432,7 @@ class ContainerAssembler
      */
     private function setupMatchers(ServiceContainer $container)
     {
-        $container->set('matchers.identity', function (ServiceContainer $c) {
-            return new Matcher\IdentityMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.comparison', function (ServiceContainer $c) {
-            return new Matcher\ComparisonMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.throwm', function (ServiceContainer $c) {
-            return new Matcher\ThrowMatcher($c->get('unwrapper'), $c->get('formatter.presenter'));
-        });
-        $container->set('matchers.type', function (ServiceContainer $c) {
-            return new Matcher\TypeMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.object_state', function (ServiceContainer $c) {
-            return new Matcher\ObjectStateMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.scalar', function (ServiceContainer $c) {
-            return new Matcher\ScalarMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.array_count', function (ServiceContainer $c) {
-            return new Matcher\ArrayCountMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.array_key', function (ServiceContainer $c) {
-            return new Matcher\ArrayKeyMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.array_contain', function (ServiceContainer $c) {
-            return new Matcher\ArrayContainMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.string_start', function (ServiceContainer $c) {
-            return new Matcher\StringStartMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.string_end', function (ServiceContainer $c) {
-            return new Matcher\StringEndMatcher($c->get('formatter.presenter'));
-        });
-        $container->set('matchers.string_regex', function (ServiceContainer $c) {
-            return new Matcher\StringRegexMatcher($c->get('formatter.presenter'));
-        });
+        $container->addDefaultExtension(new MatcherExtension());
     }
 
     /**
